move integration tests under watchOS flag
