+++ AmplifyPlugins/DataStore/Tests/DataStoreHostApp/AWSDataStorePluginIntegrationTests/Connection/DataStoreConnectionScenario1Tests.swift
@@ -32,107 +32,14 @@ import XCTest
  */
 
 class DataStoreConnectionScenario1Tests: SyncEngineIntegrationTestBase {
-
+    
     struct TestModelRegistration: AmplifyModelRegistration {
         func registerModels(registry: ModelRegistry.Type) {
             registry.register(modelType: Team1.self)
             registry.register(modelType: Project1.self)
         }
-
+        
         let version: String = "1"
-    }
-    
-
-    // MARK: - Disable Real Time Updates
-    
-    func testStartAndSync() async throws {
-        await setUp(withModels: TestModelRegistration(),
-                    logLevel: .verbose,
-                    dataStoreConfiguration: .custom(syncMaxRecords: 100, disableSubscriptions: { true }))
-        try await startAmplifyAndWaitForSync()
-    }
-    
-    func testStartAndSyncAndRestartAndSync() async throws {
-        var disabledRealTimeUpdates = true
-        let disableSubscriptions = {
-            disabledRealTimeUpdates
-        }
-        await setUp(withModels: TestModelRegistration(),
-                    logLevel: .verbose,
-                    dataStoreConfiguration: .custom(syncMaxRecords: 100, disableSubscriptions: disableSubscriptions))
-        try await startAmplifyAndWaitForSync()
-        disabledRealTimeUpdates = false
-        try await Amplify.DataStore.stop()
-        try await Amplify.DataStore.start()
-        
-        let eventReceived = expectation(description: "DataStore ready event")
-        let sink = Amplify.Hub.publisher(for: .dataStore)
-            .filter { $0.eventName == HubPayload.EventName.DataStore.ready }
-            .sink { _ in
-                eventReceived.fulfill()
-            }
-
-        try await Amplify.DataStore.start()
-
-        await fulfillment(of: [eventReceived], timeout: 10)
-    }
-    
-    func testSaveReconciled() async throws {
-        await setUp(withModels: TestModelRegistration(),
-                    logLevel: .verbose,
-                    dataStoreConfiguration: .custom(syncMaxRecords: 100, disableSubscriptions: { true }))
-        try await startAmplifyAndWaitForSync()
-        
-        let team = Team1(name: "name1")
-        let project = Project1(team: team)
-        let syncedTeamReceived = expectation(description: "received team from sync path")
-        var hubListener = Amplify.Hub.listen(to: .dataStore,
-                                             eventName: HubPayload.EventName.DataStore.syncReceived) { payload in
-            guard let mutationEvent = payload.data as? MutationEvent else {
-                XCTFail("Could not cast payload to mutation event")
-                return
-            }
-
-            if let syncedTeam = try? mutationEvent.decodeModel() as? Team1,
-               syncedTeam == team {
-                syncedTeamReceived.fulfill()
-            }
-        }
-        guard try await HubListenerTestUtilities.waitForListener(with: hubListener, timeout: 5.0) else {
-            XCTFail("Listener not registered for hub")
-            return
-        }
-
-        _ = try await Amplify.DataStore.save(team)
-        await fulfillment(of: [syncedTeamReceived], timeout: networkTimeout)
-        
-        let syncProjectReceived = expectation(description: "received project from sync path")
-        hubListener = Amplify.Hub.listen(to: .dataStore,
-                                             eventName: HubPayload.EventName.DataStore.syncReceived) { payload in
-            guard let mutationEvent = payload.data as? MutationEvent else {
-                XCTFail("Could not cast payload to mutation event")
-                return
-            }
-
-            if let syncedProject = try? mutationEvent.decodeModel() as? Project1,
-                      syncedProject == project {
-                syncProjectReceived.fulfill()
-            }
-        }
-        guard try await HubListenerTestUtilities.waitForListener(with: hubListener, timeout: 5.0) else {
-            XCTFail("Listener not registered for hub")
-            return
-        }
-        _ = try await Amplify.DataStore.save(project)
-        await fulfillment(of: [syncProjectReceived], timeout: networkTimeout)
-
-        let queriedProjectOptional = try await Amplify.DataStore.query(Project1.self, byId: project.id)
-        guard let queriedProject = queriedProjectOptional else {
-            XCTFail("Failed")
-            return
-        }
-        XCTAssertEqual(queriedProject.id, project.id)
-        XCTAssertEqual(queriedProject.team, team)
     }
     
     func testSaveTeamAndProjectSyncToCloud() async throws {
